requirements.txt
numpy>=1.22.0
pandas>=1.5.0
scikit-learn>=1.2.0
tensorflow>=2.13.0
torch>=2.0.0
pulp>=2.7.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.10.0
dash>=2.8.0
dash-bootstrap-components>=1.4.0
prefect>=2.10.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
joblib>=1.2.0
fastapi>=0.95.0
uvicorn>=0.21.0
gunicorn>=20.1.0
reportlab>=3.6.0
requests>=2.28.0 